requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==24.3.0
click==8.1.8
dark_swag==0.0.1
databases==0.9.0
fastapi==0.115.6
fastapi-utils==0.8.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
propcache==0.2.1
psutil==5.9.8
psycopg2==2.9.10
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing-inspect==0.9.0
typing_extensions==4.12.2
uvicorn==0.34.0
Werkzeug==3.1.3
yarl==1.18.3
